requirements.txt
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
cffi==1.16.0
cryptography==41.0.4
ecdsa==0.18.0
fastapi==0.103.2
Faker==19.6.2
greenlet==2.0.2
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
pymysql==1.1.0
passlib==1.7.4
protobuf==4.21.12
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
